Add neopixel debug pattern
This patch adds the use of the neopixel library for commanding the LEDS.

In terms of program flow, this patch moves the heartbeat code to run only once in setup(). Then, the patch turns on one LED at a time until all are lit, which will be helpful for assembling the LEDs and confirming that they all work.