Add use of RTClib
This commit adds RTClib as a git submodule, and it adds the include to led_clock.ino. It also runs some basic code (copied from the RTClib example) that verifies the device is able to communicate with the DS3231.

This commit also adds code to set up the serial port.