Move power LED functions to new file
This commit moves the functions that set up the power LED (polarity, softPWM) and run a heartbeat animation to a second file. After moving, the program still compiles and runs.

This commit also adds code to shut off the power LED as soon as possible after waking up. (unclear if this has a big effect)