Move input buttons to new file
This commit moves the code for initializing and reading the input buttons to a new file, and it uses the Bounce2 debouncing library to debounce the buttons. The Bounce2 library is added as a submodule. After this change, I can run the code and cause the LEDs to illuminate upon a button press.