Add LED groups to led_patterns
This commit adds several arrays of LED indices to led_patterns.ino. Each array corresponds to a set of LEDs in an interesting geometric orientation, such as the tip LEDs, the LEDs that make up the middle ring, etc. This commit also changes the main loop to call a function that illuminates one of those groups.