Add pattern switch and basic clock pattern
This commit adds a control flow in the main loop that switches between whatever pattern is being called based on a counter that is incremented by the pattern button. The control is a switch statement that calls a pattern function, passing the current time. The pattern function itself only calculates one frame of the pattern and displays it.

As an example this commit adds two pattern functions that ignore the time and just plot an interesting shape, and it adds a third pattern that is a basic clock, with hour, minute, and second indicators.

The brightness button also increments a brightness level, which is passed to the pattern function.

This commit also auto-formats the main file.

This commit also fixes a bug with the number of the pattern button #define.