Add function to get current millisecond
This commit adds the function get_now_ms, which estimates the current millisecond of time. This is useful because the RTC only has a resolution of 1 second.

The function works by being called frequently and being passed the current time from the RTC. When the current second reported by the RTC has changed, then the function marks the microcontroller clock time in milliseconds. In subsequent calls, the function returns the number of milliseconds that have elapsed since the RTC second changed.

Note: This function is not very accurate if called infrequently, and it is also not very accurate during the first second of being called.

In practice, the microcontroller clock has quite a bit of error; I found that it ran slow, with 800ish milliseconds elapsed on the microcontroller clock every time 1 second elapsed on the RTC. Therefore, I added some correction to the get_now_ms function. The correction works by measuring how slow the microcontroller clock ran in the previous second, and then it uses that factor to scale up the millisecond measurements in the current second.

After making this change, I changed the rainbow cycle pattern to use the milliseconds from this function. It looked good, but there was a visible jump once a second before I added the correction. With the correction, the rainbow cycle pattern looks as good as before.